requirements.txt
aiogram == 3.13.1
boto3 == 1.35.44
nltk == 3.9.1
googletrans == 4.0.0rc1
pydub == 0.25.1
python-dotenv == 1.0.1
langid == 1.1.6
asyncpg == 0.30.0
sqlalchemy == 2.0.36
logtail-python == 0.3.1